refactor(triggers): restructure schedule trigger into modular components
