+++ front/components/agent_builder/triggers/schedule/ScheduleEdition.tsx
@@ -1,0 +1,96 @@
+import { zodResolver } from "@hookform/resolvers/zod";
+import React, { useEffect, useMemo } from "react";
+import { useForm } from "react-hook-form";
+
+import type { AgentBuilderScheduleTriggerType } from "@app/components/agent_builder/AgentBuilderFormContext";
+import type { ScheduleFormValues } from "@app/components/agent_builder/triggers/schedule/scheduleEditionFormSchema";
+import {
+  getScheduleFormDefaultValues,
+  ScheduleFormSchema,
+} from "@app/components/agent_builder/triggers/schedule/scheduleEditionFormSchema";
+import { ScheduleEditionSheet } from "@app/components/agent_builder/triggers/schedule/ScheduleEditionSheet";
+import { FormProvider } from "@app/components/sparkle/FormProvider";
+import { useUser } from "@app/lib/swr/user";
+import type { LightWorkspaceType } from "@app/types";
+
+interface ScheduleEditionProps {
+  owner: LightWorkspaceType;
+  trigger: AgentBuilderScheduleTriggerType | null;
+  isOpen: boolean;
+  onClose: () => void;
+  onSave: (trigger: AgentBuilderScheduleTriggerType) => void;
+}
+
+export function ScheduleEdition({
+  owner,
+  trigger,
+  isOpen,
+  onClose,
+  onSave,
+}: ScheduleEditionProps) {
+  const { user } = useUser();
+
+  const isEditor = (trigger?.editor ?? user?.id) === user?.id;
+
+  const defaultValues = useMemo(
+    (): ScheduleFormValues => getScheduleFormDefaultValues(trigger),
+    [trigger]
+  );
+
+  const form = useForm<ScheduleFormValues>({
+    defaultValues,
+    resolver: zodResolver(ScheduleFormSchema),
+  });
+
+  useEffect(() => {
+    form.reset(defaultValues);
+  }, [form, defaultValues]);
+
+  const onSheetSave = async (): Promise<boolean> => {
+    await form.handleSubmit(async (values: ScheduleFormValues) => {
+      if (!user) {
+        return;
+      }
+
+      const triggerData: AgentBuilderScheduleTriggerType = {
+        sId: trigger?.sId,
+        enabled: values.enabled,
+        name: values.name.trim(),
+        kind: "schedule",
+        configuration: {
+          cron: values.cron.trim(),
+          timezone: values.timezone.trim(),
+        },
+        editor: trigger?.editor ?? user.id ?? null,
+        naturalLanguageDescription:
+          values.naturalLanguageDescription?.trim() ?? null,
+        customPrompt: values.customPrompt?.trim() ?? null,
+        editorName: trigger?.editorName ?? user.fullName ?? undefined,
+      };
+
+      onSave(triggerData);
+      onClose();
+    })();
+
+    return true;
+  };
+
+  const onCancel = () => {
+    form.reset(defaultValues);
+    onClose();
+  };
+
+  return (
+    <FormProvider form={form}>
+      <ScheduleEditionSheet
+        owner={owner}
+        trigger={trigger}
+        isOpen={isOpen}
+        onCancel={onCancel}
+        onClose={onClose}
+        onSave={onSheetSave}
+        isEditor={isEditor}
+      />
+    </FormProvider>
+  );
+}
